+++ ukat/mapping/t2.py
@@ -297,7 +297,7 @@ class T2:
                                                affine=self.affine)
                     nib.save(r2_nifti, base_path + '_r2_map.nii.gz')
                 elif result == 'mask':
-                    mask_nifti = nib.Nifti1Image(self.mask.astype(int),
+                    mask_nifti = nib.Nifti1Image(self.mask.astype(np.uint16),
                                                  affine=self.affine)
                     nib.save(mask_nifti, base_path + '_mask.nii.gz')
         else:
